+++ projeto-vida/src/components/LandPage/BannerHome/BannerHome.tsx
@@ -1,8 +1,8 @@
 import { Link } from "react-router-dom";
-import pointsImg from '../assets/images/illustrations/points.svg'
-import allUnderControlImg from '../assets/images/illustrations/all_under_control.svg'
-import '../styles/BannerHome.scss'
-import '../styles/BannerHomeResponsive.scss'
+import pointsImg from '../../../assets/images/illustrations/points.svg'
+import allUnderControlImg from '../../../assets/images/illustrations/all_under_control.svg'
+import "../../../styles/Home/BannerHome.scss"
+import "../../../styles/Home/BannerHomeResponsive.scss"
 
 export default function BannerHome() {
   return (
