+++ projeto-vida/src/components/LandPage/Header/MenuMobile.tsx
@@ -1,8 +1,8 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { Link } from 'react-router-dom'
-import back from '../assets/images/icons/ic_action_back.svg';
-import Logo from '../assets/images/icons/logo_branca.svg';
-import '../styles/MenuMobile.scss'
+import back from '../../../assets/images/icons/ic_action_back.svg';
+import Logo from '../../../assets/images/illustrations/logo_branca.svg';
+import '../../../styles/Header/MenuMobile.scss'
 export default function MenuMobile({ showMenu, setShowMenu }: { showMenu: Boolean, setShowMenu: React.Dispatch<React.SetStateAction<boolean>>; }) {
     return (
         <div id='menu-mobile' className={`${showMenu ? "active" : "inative"}`}>
