+++ projeto-vida/src/styles/BannerWarranty/BannerWarranty.scss
@@ -4,11 +4,11 @@ section#banner-warranty {
 		max-width: 179.1rem;
 		padding: 11.5rem 15.2rem;
 		border-radius: 2.4rem;
-		background: url('../assets/images/illustrations/illustra_warranty.svg')
+		background: url('../../assets/images/illustrations/illustra_warranty.svg')
 				no-repeat 100% 0 / 58.9rem,
-			url('../assets/images/illustrations/points-warranty1.svg') no-repeat
+			url('../../assets/images/illustrations/points-warranty1.svg') no-repeat
 				0% 29.7rem,
-			url('../assets/images/illustrations/points-warranty2.svg') no-repeat
+			url('../../assets/images/illustrations/points-warranty2.svg') no-repeat
 				100% 0 var(--gray-10);
 
 		h2 {
