+++ projeto-vida/src/styles/Benefits/Benefits.scss
@@ -2,11 +2,11 @@ section#benefits {
 	padding: 0.6rem 9.724rem;
 	margin-bottom: 7rem;
 	width: 100%;
-	background: url('../assets/images/illustrations/points.svg') no-repeat 0%
+	background: url('../../assets/images/illustrations/points.svg') no-repeat 0%
 			53rem,
-		url('../assets/images/illustrations/points.svg') no-repeat 100% 0rem,
-		url('../assets/images/illustrations/points.svg') no-repeat 0% 195rem,
-		url('../assets/images/illustrations/points.svg') no-repeat 100% 159rem;
+		url('../../assets/images/illustrations/points.svg') no-repeat 100% 0rem,
+		url('../../assets/images/illustrations/points.svg') no-repeat 0% 195rem,
+		url('../../assets/images/illustrations/points.svg') no-repeat 100% 159rem;
 	display: flex;
 	flex-direction: column;
 	align-items: center;
