+++ projeto-vida/src/styles/Benefits/BenefitsResponsive.scss
@@ -1,4 +1,4 @@
-@import "globals";
+@import "../globals.scss";
 section#benefits {
 	@media screen and (max-width: $large) {
 		div.benefit {
