+++ projeto-vida/src/styles/Courses/CoursesResponsive.scss
@@ -1,4 +1,4 @@
-@import "globals";
+@import "../globals";
 section#carousel {
 	@media screen and (max-width: $extra-large) {
 		padding: 11.8rem;
