+++ projeto-vida/src/styles/EventSection/BannerEvents.scss
@@ -1,7 +1,7 @@
 section#banner-event {
 	margin: 7rem 6.5rem;
-	background: url("../assets/images/illustrations/points-events.svg") no-repeat 1% 5.7rem,
-		url("../assets/images/illustrations/points-events.svg") no-repeat 100% 100% var(--gray-10);
+	background: url("../../assets/images/illustrations/points-events.svg") no-repeat 1% 5.7rem,
+		url("../../assets/images/illustrations/points-events.svg") no-repeat 100% 100% var(--gray-10);
 	max-width: 179.1rem;
 	padding: 8.5rem 15.2rem;
 	border-radius: 2.4rem;
