+++ projeto-vida/src/styles/EventSection/EventCard.scss
