+++ projeto-vida/src/styles/EventSection/EventCardResponsive.scss
@@ -1,4 +1,4 @@
-@import "globals";
+@import "../globals";
 div#event-card-container {
     @media screen and(max-width: $extra-large) {
         #event-card {
