+++ projeto-vida/src/styles/Experience/ExperienceResponsive.scss
@@ -1,4 +1,4 @@
-@import "globals";
+@import "../globals";
 section#experience {
 	//1536px
 	@media screen and (max-width: $extra-large) {
