+++ projeto-vida/src/styles/Footer/Footer.scss
@@ -5,8 +5,8 @@ footer {
 	flex-direction: column;
 	justify-content: center;
 	align-items: center;
-	background: url("../assets/images/illustrations/points-events.svg") no-repeat 1% 1rem,
-		url("../assets/images/illustrations/points-events.svg") no-repeat 99% 90% var(--gray-1);
+	background: url("../../assets/images/illustrations/points-events.svg") no-repeat 1% 1rem,
+		url("../../assets/images/illustrations/points-events.svg") no-repeat 99% 90% var(--gray-1);
 
 	p {
 		font-family: "aller_lt";
