+++ projeto-vida/src/styles/Header/HeaderResponsive.scss
@@ -1,4 +1,4 @@
-@import "globals";
+@import "../globals";
 header {
 	@media screen and (max-width: $extra-large) {
 		height: 17rem;
