+++ projeto-vida/src/styles/Header/MenuMobile.scss
@@ -1,4 +1,4 @@
-@import "globals";
+@import "../globals";
 #menu-mobile {
     position: absolute;
     top: 0;
