+++ projeto-vida/src/styles/Home/BannerHomeResponsive.scss
@@ -1,4 +1,4 @@
-@import "globals";
+@import "../globals";
 #banner-home {
 	#content-banner-home {
 		@media screen and (max-width: $extra-large) {
